+++ tests/features/test_smarterdb.py
@@ -116,7 +116,7 @@ class BreedTestCase(MongoMockMixin, unittest.TestCase):
 class VariantMixin():
     @classmethod
     def setUpClass(cls):
-        with open(DATA_DIR / "affy_variants.json") as handle:
+        with open(DATA_DIR / "variants.json") as handle:
             cls.data = json.load(handle)[0]
 
         with open(DATA_DIR / "affy_variants.json") as handle: